Update README.md
Added Info 
## Legacy

"Code Geass" has achieved classic status in the mecha and psychological thriller genres. It received critical acclaim for its gripping storyline, well-developed characters, and unexpected twists that captivated viewers. The conclusion of the first season and its sequel, "Code Geass: Lelouch of the Rebellion R2," are particularly noteworthy for their emotional impact and resolution of the complex narrative.